# openapi-oficina.yaml
openapi: 3.0.3
info:
  title: API Oficina Mecânica
  version: "v1"
  description: |
    API em Python (Flask) para gestão de autenticação de usuários (JWT), busca externa
    de produtos (token de serviço) e carrinho por usuário. Documentação em PT-BR.
  contact:
    name: Suporte da API
    email: suporte@example.com
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Desenvolvimento local
...
security:
  - bearerAuth: []
paths:
...
  /montadoras:
    get:
      tags: [Busca]
      summary: Listar montadoras (catálogo externo)
      description: Não exige JWT de usuário; usa token de serviço interno.
      responses:
        '200': { description: Lista de montadoras }
        '500': { $ref: '#/components/responses/Resp500' }
  /familias:
    get:
      tags: [Busca]
      summary: Listar famílias
      responses:
        '200': { description: Lista de famílias }
        '500': { $ref: '#/components/responses/Resp500' }
  /familias/{familia_id}/subfamilias:
    get:
      tags: [Busca]
      summary: Listar subfamílias por família
      parameters:
        - in: path
          name: familia_id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200': { description: Lista de subfamílias }
        '404': { $ref: '#/components/responses/Resp404' }
  /autocomplete:
    get:
      tags: [Busca]
      summary: Autocomplete ao vivo
      parameters:
        - in: query
          name: prefix
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Sugestões
          content:
            application/json:
              schema:
                type: object
                properties:
                  sugestoes:
                    type: array
                    items: { type: string }
  /facetas-produto:
    get:
      tags: [Busca]
      summary: Facetas (subprodutos/marcas) para filtro
      parameters:
        - in: query
          name: produto_nome
          schema: { type: string }
        - in: query
          name: familia_nome
          schema: { type: string }
        - in: query
          name: familia_id
          schema: { type: integer }
        - in: query
          name: subfamilia_id
          schema: { type: integer }
        - in: query
          name: placa
          schema: { type: string }
      responses:
        '200': { description: Facetas calculadas }
        '400': { $ref: '#/components/responses/Resp400' }
  /pesquisar:
    get:
      tags: [Busca]
      summary: Pesquisar produtos
      description: Por termo ou por família/subfamília; suporta placa, marca e ordenação.
      parameters:
        - in: query
          name: termo
          schema: { type: string }
        - in: query
          name: familia_id
          schema: { type: integer }
        - in: query
          name: familia_nome
          schema: { type: string }
        - in: query
          name: subfamilia_id
          schema: { type: integer }
        - in: query
          name: marca
          schema: { type: string }
        - in: query
          name: placa
          schema: { type: string }
        - in: query
          name: ordenar_por
          schema: { type: string, example: "preco_final" }
        - in: query
          name: ordem
          schema: { type: string, enum: [asc, desc], default: asc }
        - in: query
          name: pagina
          schema: { type: integer, default: 1 }
        - in: query
          name: itens_por_pagina
          schema: { type: integer, default: 20 }
      responses:
        '200': { description: Resultados paginados }
        '400': { $ref: '#/components/responses/Resp400' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Resp400: { description: Requisição inválida }
...
    Resp404: { description: Não encontrado }
...
    Resp500: { description: Erro interno }
...